requirements.txt
aldjemy==0.3.51
Django==1.6.5
django-extensions==1.5.0
django-webtest==1.7.7
factory-boy==2.4.1
SQLAlchemy==0.9.8
pytest==2.6.4
pytest-cov==1.8.1
pytest-django==2.8.0
pytest-pep8==1.0.6
fake-factory==0.4.2
flake8==2.3.0
South==1.0.2
